+++ plugins/woocommerce-admin/client/dashboard/task-list/tasks/products.js
@@ -9,7 +9,7 @@ import { Component, Fragment } from '@wordpress/element';
  * WooCommerce dependencies
  */
 import { Card, List } from '@woocommerce/components';
-import { getAdminLink } from '@woocommerce/navigation';
+import { getAdminLink } from '@woocommerce/wc-admin-settings';
 
 /**
  * Internal dependencies