+++ plugins/woocommerce-admin/client/wp-admin-scripts/onboarding-homepage-notice/index.js
@@ -10,7 +10,7 @@ import domReady from '@wordpress/dom-ready';
 /**
  * WooCommerce dependencies
  */
-import { getAdminLink } from '@woocommerce/navigation';
+import { getAdminLink } from '@woocommerce/wc-admin-settings';
 
 /**
  * Internal dependencies