+++ plugins/woocommerce-admin/packages/components/src/link/index.js
@@ -9,7 +9,7 @@ import { partial } from 'lodash';
 /**
  * WooCommerce dependencies
  */
-import { getAdminLink, getHistory } from '@woocommerce/navigation';
+import { getHistory } from '@woocommerce/navigation';
 
 /**
  * Use `Link` to create a link to another resource. It accepts a type to automatically
@@ -36,13 +36,6 @@ class Link extends Component {
 	render() {
 		const { children, href, type, ...props } = this.props;
 
-		let path;
-		if ( 'wp-admin' === type ) {
-			path = getAdminLink( href );
-		} else {
-			path = href;
-		}
-
 		const passProps = {
 			...props,
 			'data-link-type': type,
@@ -53,7 +46,7 @@ class Link extends Component {
 		}
 
 		return (
-			<a href={ path } { ...passProps }>
+			<a href={ href } { ...passProps }>
 				{ children }
 			</a>
 		);