+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Export the `<CompareButton />` component.
 - Add `<TextControl />` component.
 - Require `currency` prop in `<AdvancedFilters />` component.
+- Remove call to `getAdminLink()` inside the `<Link />` component.
 
 # 4.0.0
 - Added a new `<ScrollTo />` component.